extract each lint to a separate module
